requirements.txt
pandas
numpy
numpy-financial
pyyaml
streamlit
plotly
